+++ src/main/java/org/sopt/utils/PostUtil.java
@@ -1,4 +1,4 @@
-package org.sopt.util;
+package org.sopt.utils;
 
 public class PostUtil {
     private static int postId = 0;
